area light concept and shadows rays (to center). multiple samples: read on.
multiple samples are based on mixed-up surface information due to
re-using the ray data for the shadow trace. thus, we have to take care
to keep the original surface information for further shadow samples.

this might be made optional as this would incur overhead for path
tracing, but otoh, this is still future work anyway.