last commit introduced a bug, read on.
reset of gpu intersections war supposed to be handled by gpu_raytracer in the
iteratred_tracer. the iterated_tracer, however, generally derives from
basic_raytracer, so the empty default implementation was used, and intersection
data therefore uninitialized...

deriving from gpu_raytracer introduces the sibling tracers at this level, which
should not be a problem, but was deferred anyhow.